code updated
calculated white spots in 2018 as those newsrooms which existed for the entire duration of 2018. On 29th December 2018, 10 newsrooms changed owner. In the MedieStudier database, this was split into two observations, both of which were excluded in my filter function. To include at least one, I changed the filter function to all those newsrooms which existed up until 28th december included, so I would not capture an inaccurate number of white spots.